Centralise static GraphQL types outside of makeAugmentedSchema
