+++ packages/graphql/src/schema/types/input-objects/CartesianPointInput.ts
@@ -17,10 +17,19 @@
  * limitations under the License.
  */
 
-export { default as BigInt } from "./BigInt";
-export { default as DateTime } from "./DateTime";
-export { default as Date } from "./Date";
-export { default as Duration } from "./Duration";
-export { default as LocalDateTime } from "./LocalDateTime";
-export { default as Time } from "./Time";
-export { default as LocalTime } from "./LocalTime";
+import { GraphQLFloat, GraphQLInputObjectType, GraphQLNonNull } from "graphql";
+
+export const CartesianPointInput = new GraphQLInputObjectType({
+    name: "CartesianPointInput",
+    fields: {
+        x: {
+            type: new GraphQLNonNull(GraphQLFloat),
+        },
+        y: {
+            type: new GraphQLNonNull(GraphQLFloat),
+        },
+        z: {
+            type: GraphQLFloat,
+        },
+    },
+});
