+++ packages/graphql/src/schema/types/input-objects/QueryOptions.ts
@@ -1,0 +1,13 @@
+import { GraphQLInputObjectType, GraphQLInt } from "graphql";
+
+export const QueryOptions = new GraphQLInputObjectType({
+    name: "QueryOptions",
+    fields: {
+        offset: {
+            type: GraphQLInt,
+        },
+        limit: {
+            type: GraphQLInt,
+        },
+    },
+});
