+++ packages/graphql/src/schema/types/objects/CreateInfo.ts
@@ -1,0 +1,16 @@
+import { GraphQLInt, GraphQLNonNull, GraphQLObjectType, GraphQLString } from "graphql";
+
+export const CreateInfo = new GraphQLObjectType({
+    name: "CreateInfo",
+    fields: {
+        bookmark: {
+            type: GraphQLString,
+        },
+        nodesCreated: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+        relationshipsCreated: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+    },
+});
