+++ packages/graphql/src/schema/types/objects/DeleteInfo.ts
@@ -1,0 +1,16 @@
+import { GraphQLInt, GraphQLNonNull, GraphQLObjectType, GraphQLString } from "graphql";
+
+export const DeleteInfo = new GraphQLObjectType({
+    name: "DeleteInfo",
+    fields: {
+        bookmark: {
+            type: GraphQLString,
+        },
+        nodesDeleted: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+        relationshipsDeleted: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+    },
+});
