+++ packages/graphql/src/schema/types/objects/UpdateInfo.ts
@@ -1,0 +1,22 @@
+import { GraphQLInt, GraphQLNonNull, GraphQLObjectType, GraphQLString } from "graphql";
+
+export const UpdateInfo = new GraphQLObjectType({
+    name: "UpdateInfo",
+    fields: {
+        bookmark: {
+            type: GraphQLString,
+        },
+        nodesCreated: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+        nodesDeleted: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+        relationshipsCreated: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+        relationshipsDeleted: {
+            type: new GraphQLNonNull(GraphQLInt),
+        },
+    },
+});
