+++ packages/graphql/src/schema/types/scalars/index.ts
@@ -1,0 +1,7 @@
+export { BigInt } from "./BigInt";
+export { Date } from "./Date";
+export { DateTime } from "./DateTime";
+export { Duration } from "./Duration";
+export { LocalDateTime } from "./LocalDateTime";
+export { LocalTime } from "./LocalTime";
+export { Time } from "./Time";
