+++ packages/graphql/tests/integration/types/localtime.int.test.ts
@@ -23,7 +23,7 @@ import neo4jDriver, { Driver } from "neo4j-driver";
 import { generate } from "randomstring";
 import neo4j from "../neo4j";
 import { Neo4jGraphQL } from "../../../src/classes";
-import { parseLocalTime } from "../../../src/schema/scalars/LocalTime";
+import { parseLocalTime } from "../../../src/schema/types/scalars/LocalTime";
 
 describe("LocalTime", () => {
     let driver: Driver;
